Refactoring to be able to use template class/methods even if they are not initiated
